fix: the selection should be cleared if selecting child node
